Reformat to reuse common code
