+++ rs_workflows/on_demand/sentinel3/s3_l0.py
@@ -24,6 +24,10 @@ from rs_workflows.flow_utils import (
 )
 from rs_workflows.on_demand.common.l0_last_steps import process_l0_last_steps
 from rs_workflows.on_demand.common.types import Level0FlowParams
+from rs_workflows.on_demand.sentinel3.s3_processing_utils import (
+    build_olci_l1_input_products,
+    read_s3_orchestration_settings,
+)
 
 S3_L0_RESULT_STORAGE = "local-file-system/s3-processing-shared-results"
 S3_L0_PRODUCTS_READY_EVENT = "rs-python.s3-l0.products-ready"
@@ -63,6 +67,11 @@ async def process_s3l0(
         input_products=input_products,
         verbose=verbose,
     )
+    orchestration_settings = await read_s3_orchestration_settings()
+    l1_input_products = build_olci_l1_input_products(
+        products,
+        orchestration_settings.s3_l0_output_collection,
+    )
 
     flow_run_id = str(runtime.flow_run.id or "unknown")
     emitted_event = emit_event(
@@ -83,6 +92,7 @@ async def process_s3l0(
             "session_id": session,
             "owner_identifier": resolved_flow_params.owner_identifier,
             "products": products,
+            "input_products": l1_input_products,
         },
     )
     logger = get_run_logger()
